+++ kernel/include/interrupt/timer.h
@@ -7,9 +7,25 @@
 #define TIMER_FREQUENCY_MAX 1193180
 #define TIMER_FREQUENCY_MIN 2
 
-void init_timer(uint32_t frequency, interrupt_handler_t timer_callback);
+#define TIMER_FRE 10000 // testing frequency, in republish version, it shouble be 1193180
+
+
+typedef void (*timer_callback)(void);
+
+// kernel timer
+typedef struct timer {
+    timer_callback callback_func;  // call back function
+    uint32_t interval;
+    uint32_t elapse;
+    struct timer *next;            // linked list next 
+} timer_t;
 
 
 
+
+
+void init_timer();
+void register_timer(timer_callback callback, uint32_t frequency);
+void unregister_timer(timer_t* timer);
 
 #endif