+++ kernel/kernel/pmm.c
@@ -1,6 +1,7 @@
 #include "kernel/pmm.h" // show_memory_map
 #include <stdio.h> // printf
 #include "kernel/vmm.h"
+#include "kernel/math.h"
 normal_mem_t normal_mem; // global variable storing normal memory zone information
 
 /**
@@ -21,11 +22,6 @@ normal_mem_t normal_mem; // global variable storing normal memory zone informati
  *      - uint32_t frame_free : free the pages allocated by frame_alloc
  * */
 
-
-
-static inline uint32_t is_power_of_2 (uint32_t n);
-static inline uint32_t next_power_of_2 (uint32_t n);
-static inline uint32_t power_of_2 (uint32_t n);
 
 // buddy_system
 
@@ -277,30 +273,3 @@ inline uint32_t buddy_system_free_frame_page_block(uint32_t page_index) {
     return 1;
 }
 
-
-
-inline uint32_t is_power_of_2 (uint32_t n) {
-    return !((n-1) & n) && n;
-}
-
-
-inline uint32_t next_power_of_2 (uint32_t n) {
-    if(is_power_of_2(n))
-        return n;
-    n |= n >> 1;
-    n |= n >> 2;
-    n |= n >> 4;
-    n |= n >> 8;
-    n |= n >> 16;
-
-    return n + 1;
-}
-
-inline uint32_t power_of_2 (uint32_t n) {
-    uint32_t i = 0;
-    while(! (n & (uint32_t)1)) {
-        n = n >> 1;
-        ++i;
-    }
-    return i;
-}