+++ kernel/kernel/schedule.c
@@ -8,7 +8,7 @@ extern task_list_node_t* current_task;
 
 static switch_context(context_t* cur, context_t* next);
 
-void task_schedule() {
+timer_callback task_schedule() {
     if(current_task) {
         task_list_node_t* next = current_task -> next;
         if(next != current_task) {
@@ -21,5 +21,5 @@ void task_schedule() {
 
 void init_schedule() {
     init_task();
-    init_timer(TASK_SWITCH_FREQUENCY, &task_schedule);
+    register_timer(TASK_SWITCH_FREQUENCY, &task_schedule);
 }